Restructuring the application as a namespace package. RE:#1
