+++ exe/root.spec
@@ -8,7 +8,7 @@ from PyInstaller.compat import is_win, is_darwin
 
 block_cipher = None
 
-path_extension = []
+path_extension = ['rootcode']
 conda_env = os.environ['CONDA_PREFIX']
 
 if is_win:
@@ -19,7 +19,7 @@ else:
 # Add the root_ui directory to the extended path.
 path_extension.insert(0, os.path.abspath('..'))
 
-a = Analysis([os.path.join(os.getcwd(), 'rootcode', 'root.py')],  # Assume we're building from the project root
+a = Analysis([os.path.join(os.getcwd(), 'natcap', 'root', 'root.py')],  # Assume we're building from the project root
              pathex=path_extension,
              binaries=None,
              datas=[('qt.conf', '.'), proj_datas],
