+++ natcap/__init__.py
@@ -1,0 +1,4 @@
+# this is a namespace package
+
+import pkg_resources
+pkg_resources.declare_namespace(__name__)
